talkitty-python-gateway-microservice/requirements.txt
# Core Django and DRF
Django==4.2.7
djangorestframework==3.14.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP and Networking
requests==2.31.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Channels for WebSocket support (optional)
channels==4.0.0

# Production and Deployment
gunicorn==21.2.0
psycopg2-binary==2.9.9

# Development and Debugging
ipython==8.17.2

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Additional Utilities
uuid==1.30